+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/synthesis/SynthesisScreen.java
@@ -148,8 +148,6 @@ public class SynthesisScreen extends MenuBackground implements IPlayerDataReques
 
 	@Override
 	public void render(@NotNull GuiGraphics gui, int mouseX, int mouseY, float partialTicks) {
-		
-		//fill(125, ((-140 / 16) + 75) + 10, 200, ((-140 / 16) + 75) + 20, 0xFFFFFF);
 		super.render(gui, mouseX, mouseY, partialTicks);
 	}